change to one folder structure
