+++ meta/squads/squad_2025-01-25.csv
@@ -1,0 +1,16 @@
+,web_name,now_cost,player_position,team_name,chance_of_playing_next_round,mean_ppg_7
+3,Pickford,50,GKP,Everton,,4.14
+9,Dúbravka,43,GKP,Newcastle,100.0,5.14
+0,Kerkez,48,DEF,Bournemouth,100.0,5.57
+2,Huijsen,44,DEF,Bournemouth,,5.57
+4,Alexander-Arnold,73,DEF,Liverpool,100.0,5.57
+10,Hall,51,DEF,Newcastle,,5.57
+12,Aina,54,DEF,Nott'm Forest,100.0,6.0
+8,Amad,56,MID,Man Utd,100.0,5.29
+13,Elanga,53,MID,Nott'm Forest,100.0,6.14
+1,Kluivert,56,MID,Bournemouth,100.0,7.71
+6,M.Salah,137,MID,Liverpool,,9.29
+7,Foden,93,MID,Man City,100.0,9.43
+5,Gakpo,75,FWD,Liverpool,,5.86
+14,Wood,71,FWD,Nott'm Forest,100.0,6.71
+11,Isak,95,FWD,Newcastle,100.0,9.57
