+++ src/select_fpl_squad.py
@@ -1,6 +1,9 @@
 import pandas as pd
 from pulp import LpProblem, LpVariable, LpMaximize, lpSum, LpStatus, PULP_CBC_CMD
 
+from data_sourcing import fetch_fpl_data
+from feature_engineering import engineer_features
+
 import logging
 logging.basicConfig(level=logging.INFO)
 
@@ -10,6 +13,7 @@ def select_fpl_squad(df,
                      num_defs=5,
                      num_mids=5,
                      num_atts=3,
+                     num_mngs=0,
                      max_value=1000):
     """
     Selects the optimal 15-man squad based on the given constraints.
@@ -27,6 +31,18 @@ def select_fpl_squad(df,
         - selected_squad: pandas dataframe containing the selected squad
     """
     logging.info("Selecting optimal FPL squad.")
+
+    cols_to_keep = [
+        'id_player',
+        'web_name',
+        'now_cost',
+        'player_position',
+        'team_name',
+        'chance_of_playing_next_round',
+        metric,
+    ]
+
+    df = df[cols_to_keep].drop_duplicates()
 
     # Reset index to ensure it ranges from 0 to N-1
     df = df.reset_index(drop=True)
@@ -50,7 +66,7 @@ def select_fpl_squad(df,
     prob += lpSum(df.loc[i, 'now_cost'] * x[i] for i in player_ids) <= max_value, "Total_value"
 
     # Position constraints
-    positions = {'GKP': num_gks, 'DEF': num_defs, 'MID': num_mids, 'FWD': num_atts}
+    positions = {'GKP': num_gks, 'DEF': num_defs, 'MID': num_mids, 'FWD': num_atts, 'MNG': num_mngs}
     for pos, count in positions.items():
         pos_ids = df[df['player_position'] == pos].index.tolist()
         prob += lpSum(x[i] for i in pos_ids) == count, f"Total_{pos.upper()}"
@@ -80,3 +96,12 @@ def select_fpl_squad(df,
     selected_squad = df.loc[selected_ids].reset_index(drop=True)
 
     return selected_squad
+
+if __name__ == '__main__':
+    num_gws = 7
+    metric = f'mean_ppg_{num_gws}'
+
+    df = fetch_fpl_data()
+    df = engineer_features(df, num_gws)
+    squad = select_fpl_squad(df, metric)
+    print(squad.sort_values(by='player_position'))
